requirements.txt
numpy
pandas
seaborn
fastapi
aiofiles
uvicorn
pyserial
py-sds011
python-dotenv
pymongo
